Move resource type configuration to modes
