+++ modes/safe-resource-type.js
